+++ CHANGELOG.rst
@@ -2,13 +2,14 @@
 Changelog history
 ==================
 
-Esmlab v2019.3. (2019-03)
+Esmlab v2019.4. (2019-04)
 ==============================
 
 Features
 --------
 
 - Enable computing significance metrics (p-value) for ``weighted_corr`` (:pr:`90`) `Riley Brady`_
+- Add ``compute_mon_mean`` function to compute monthly averages (:pr:`110`) `Alper Altuntas`_
 
 
 Esmlab v2019.3.16 (2019-03-16)
@@ -88,7 +89,7 @@ Esmlab v2019.2.0 (2019-02-02)
 
 
 
-
+.. _`Alper Altuntas`: https://github.com/alperaltuntas
 .. _`Anderson Banihirwe`: https://github.com/andersy005
 .. _`Matthew Long`: https://github.com/matt-long
 .. _`Michael Levy`: https://github.com/mnlevy1981